Detect Google Adsense
Google Adsense is a program run by Google that allows publishers of
content sites to serve automatic advertisement. Unfortunately, it means
tracking and breaching privacy of our users.

Moreover using Google Adsense in local webpage copy is a violation
of Google Adsense term of use and thus a serious bug that render
not distribuable this package.